+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All changes to this project will be documented in this file.
 ## [1.0.8] - 2022-04-27
 - Create a AbstractUploader
 - Add origin header
+- Add the possibility to provide a refresh token
 
 ## [1.0.7] - 2022-04-26
 - Don't retry on 401 errors